Refactor(web): Migrate RemoveSpendingLimitFlow
